linkProviders in own module
