+++ lib/common.js
@@ -4,103 +4,13 @@ import { existsSync } from 'node:fs';
 import { stat, readFile, writeFile, readdir } from 'node:fs/promises';
 import https from 'node:https';
 import path from 'node:path';
-import queryString from 'node:querystring';
-import url from 'node:url';
 import { unified } from 'unified';
 import remarkParse from 'remark-parse';
 import remarkRehype from 'remark-rehype';
 import rehypeStringify from 'rehype-stringify';
 import sharp from 'sharp';
 import yaml from 'js-yaml';
-
-
-const linkProviders = [{
-  id: 'vimeo',
-  test: urlObj => urlObj.hostname === 'vimeo.com',
-  // getObject: (linkNode, urlObj) => ({
-  //   type: 'video',
-  //   provider: 'vimeo',
-  //   href: linkNode.url,
-  //   id: urlObj.pathname.slice(1),
-  // }),
-  getHTML: (urlObj) => {
-    const videoId = urlObj.pathname.slice(1);
-    const qs = queryString.encode({
-      title: 0,
-      byline: 0,
-      portrait: 0,
-      ...queryString.decode(urlObj.query),
-    });
-    const src = `https://player.vimeo.com/video/${videoId}?${qs}`;
-
-    const attributes = {
-      width: 640,
-      height: 360,
-      src,
-      frameBorder: '0',
-      scrolling: 'no',
-      allowfullscreen: 'true',
-      mozallowfullscreen: 'true',
-      webkitallowfullscreen: 'true',
-    };
-
-    return `<iframe ${Object.entries(attributes).map(([key, value]) => `${key}="${value}"`).join(' ')}></iframe>`;
-  },
-},
-{
-  id: 'youtube',
-  test: (urlObj, qs) => (
-    urlObj.hostname === 'youtu.be'
-    || (
-      urlObj.hostname === 'www.youtube.com'
-      && urlObj.pathname === '/watch'
-      && qs.v
-    )
-  ),
-  getObject: (a, urlObj, qs) => ({
-    type: 'video',
-    provider: 'youtube',
-    href: a.href,
-    id: urlObj.hostname === 'youtu.be' ? urlObj.pathname.slice(1) : qs.v,
-  }),
-  getHTML: (urlObj, { v, ...qs }) => {
-    const videoId = urlObj.hostname === 'youtu.be' ? urlObj.pathname.slice(1) : v;
-    const params = queryString.stringify({ ...qs, autoplay: 0 });
-    const attributes = {
-      width: 640,
-      height: 360,
-      src: `https://www.youtube.com/embed/${videoId}?${params}`,
-      frameBorder: '0',
-      gesture: 'media',
-      allow: 'encrypted-media',
-      allowfullscreen: 'true',
-    };
-    return `<iframe ${Object.entries(attributes).map(([key, value]) => `${key}="${value}"`).join(' ')}></iframe>`;
-  },
-}];
-
-const transformLink = (node) => {
-  if (node.type !== 'link') {
-    return node;
-  }
-  const urlObj = url.parse(node.url);
-  const qs = queryString.parse(urlObj.query);
-  const provider = linkProviders.find(item => item.test(urlObj, qs));
-
-  if (!provider) {
-    return node;
-  }
-
-  // at this point the nodes are of mdast markdown
-  // and if we want to transform to an iframe
-  // we can only do node type html https://github.com/syntax-tree/mdast#nodes
-  // example {type: 'html', value: '<div>'}
-
-  // TODO use children text for alt tag ?
-  // TODO put in iframe container https://github.com/Laboratoria/curriculum-parser/blob/main/lib/common.js#L315
-  const htmlNode = { ...node, type: 'html', value: provider.getHTML(urlObj, qs) };
-  return htmlNode;
-};
+import { parseLinks } from './linkProviders.js';
 
 export const parser = unified()
   .use(remarkParse)
@@ -294,18 +204,6 @@ export const parseReadmes = async (dir, langs, type, fn) => {
 
         const [metaText, contentText] = splitMetaAndContent(trimmed);
         const rootNode = parser.parse(contentText);
-
-        // eslint-disable-next-line no-unused-vars
-        const parseLinks = (node) => {
-          if (node.type === 'link') {
-            const linkProps = transformLink(node);
-            Object.assign(node, linkProps);
-          }
-          if (node.children?.length) {
-            return ({ ...node, children: node.children.map(n => parseLinks(n)) });
-          }
-          return node;
-        };
 
         const meta = (
           !metaText
